- Ensure DataTable columns are correctly typed - Break commands into command processor
