+++ applications/lokqlDx/MainWindow.xaml.cs
@@ -10,6 +10,7 @@ using KustoLoco.Core.Settings;
 using KustoLoco.FileFormats;
 using KustoLoco.Rendering;
 using Lokql.Engine;
+using Lokql.Engine.Commands;
 using Microsoft.Win32;
 using NotNullStrings;
 using ZstdSharp.Unsafe;
@@ -34,7 +35,8 @@ public partial class MainWindow : Window
         var settings = new KustoSettingsProvider();
         _workspaceManager = new WorkspaceManager(settings);
         var loader = new StandardFormatAdaptor(settings,_console);
-        _explorer = new InteractiveTableExplorer(_console, loader, settings);
+        var cp = new CommandProcessor();
+        _explorer = new InteractiveTableExplorer(_console, loader, settings,cp);
         
     }
 
@@ -54,7 +56,7 @@ public partial class MainWindow : Window
         //that the UI thread is not blocked for reports generated by
         //the engine
       
-        await Task.Run(async () => await _explorer.RunInput(query, false));
+        await Task.Run(async () => await _explorer.RunInput(query));
         var  result = _explorer._prevResult;
         //if there are no results leave the previously rendered results in place
         if (result.RowCount != 0)
@@ -65,7 +67,7 @@ public partial class MainWindow : Window
                 //annoying we have to do this, but it's the only way to get the webview to render
                 await webview.EnsureCoreWebView2Async();
                 //generate the html and display it
-                var renderer = new KustoResultRenderer(_explorer._settings);
+                var renderer = new KustoResultRenderer(_explorer.Settings);
                 var html = renderer.RenderToHtml(result);
                 try
                 {
@@ -104,7 +106,7 @@ public partial class MainWindow : Window
         Editor.SetText(_workspaceManager.UserText);
         var settings = _workspaceManager.Settings;
         var loader = new StandardFormatAdaptor(settings,_console);
-        _explorer = new InteractiveTableExplorer(_console, loader, settings);
+        _explorer = new InteractiveTableExplorer(_console, loader, settings,new CommandProcessor());
         UpdateFontSize();
         Title = $"LokqlDX - {_workspaceManager.Path}";
     }
