+++ libraries/lokql-engine/Commands/ListSettingDefinitionsCommand.cs
@@ -1,0 +1,25 @@
+﻿using CommandLine;
+
+namespace Lokql.Engine.Commands;
+
+public static class ListSettingDefinitionsCommand
+{
+    internal static Task RunAsync(InteractiveTableExplorer exp, Options o)
+    {
+        var defs = exp.Settings.GetDefinitions()
+            .Where(d => d.Name.Contains(o.Match, StringComparison.InvariantCultureIgnoreCase))
+            .OrderBy(d => d.Name);
+
+        var str = Tabulator.Tabulate(defs, "Name|Description|Default", d => d.Name, d => d.Description,
+            d => d.DefaultValue);
+        exp.Info(str);
+        return Task.CompletedTask;
+    }
+
+    [Verb("settingdefinitions", HelpText = "lists all setting definitions")]
+    internal class Options
+    {
+        [Value(0, HelpText = "match substring", Required = false)]
+        public string Match { get; set; } = string.Empty;
+    }
+}
